Refactoring. (Page class)
