+++ src/main/kotlin/com/example/quo_digital_demo/services/AuthorsService.kt
@@ -6,12 +6,7 @@ import java.util.UUID
 
 @Service
 interface AuthorsService {
-
-    data class Page(
-        val authors: List<Author>,
-        val numberOfAllAuthors: Int)
-
-    fun findPage(pageNum: Int, itemsPerPage: Int): Page
+    fun findPage(pageNum: Int, itemsPerPage: Int): Page<Author>
     fun findById(id: UUID): Author?
     fun create(fullName: String): Author
     fun update(author: Author): Int
