+++ src/main/kotlin/com/example/quo_digital_demo/services/BooksService.kt
@@ -6,13 +6,8 @@ import java.util.UUID
 
 @Service
 interface BooksService {
-
-    data class Page(
-        val books: List<Book>,
-        val numberOfAllBooks: Int)
-
-    fun findPageByAuthorId(authorId: UUID, pageNum: Int, itemsPerPage: Int): Page
-    fun findPage(pageNum: Int, itemsPerPage: Int): Page
+    fun findPageByAuthorId(authorId: UUID, pageNum: Int, itemsPerPage: Int): Page<Book>
+    fun findPage(pageNum: Int, itemsPerPage: Int): Page<Book>
     fun findById(id: UUID): Book?
     fun create(title: String, authorIds: List<UUID>): Book
     fun update(book: Book): Int
