+++ src/main/kotlin/com/example/quo_digital_demo/services/BooksServiceImpl.kt
@@ -2,21 +2,20 @@ package com.example.quo_digital_demo.services
 
 import com.example.quo_digital_demo.models.Book
 import com.example.quo_digital_demo.repositories.BooksRepository
-import com.example.quo_digital_demo.services.BooksService.Page
 import org.springframework.stereotype.Service
 import java.util.UUID
 
 @Service
 class BooksServiceImpl(private val booksRepository: BooksRepository) : BooksService {
 
-    override fun findPageByAuthorId(authorId: UUID, pageNum: Int, itemsPerPage: Int): Page {
+    override fun findPageByAuthorId(authorId: UUID, pageNum: Int, itemsPerPage: Int): Page<Book> {
         val offset = (pageNum - 1) * itemsPerPage
         val books = booksRepository.findByAuthorIdWithOffsetLimit(authorId, offset, itemsPerPage)
         val allNum = booksRepository.countAllByAuthorId(authorId)
         return Page(books, allNum)
     }
 
-    override fun findPage(pageNum: Int, itemsPerPage: Int): Page {
+    override fun findPage(pageNum: Int, itemsPerPage: Int): Page<Book> {
         val offset = (pageNum - 1) * itemsPerPage
         val books = booksRepository.findWithOffsetLimit(offset, itemsPerPage)
         val allNum = booksRepository.countAll()
