+++ src/main/kotlin/com/example/quo_digital_demo/services/Page.kt
@@ -1,0 +1,5 @@
+package com.example.quo_digital_demo.services
+
+data class Page<T>(
+    val items: List<T>,
+    val itemsTotalNum: Int)
